+++ config/settings_model.go
@@ -1,0 +1,7 @@
+package config
+
+type SettingsModel struct {
+	APIKey     string
+	ServerPort string
+	FilesDir   string
+}
