+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/mnmonherdene1234/files-gin-go
 
-go 1.22
+go 1.23.2
 
 require (
 	github.com/gin-contrib/cors v1.7.2
